Refactor shim diagnostics attachment ownership in tree wiring
